razorpay integrated

+++ Expense Tracker/controllers/controllers.js
@@ -1,8 +1,11 @@
+const { where } = require("sequelize");
 const Expense = require("../models/expense");
 const User = require("../models/users");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
-
+const Razorpay = require("razorpay");
+const Order = require("../models/order");
+require('dotenv').config();
 
 
 exports.signUp = async (req, res, next) => {
@@ -43,8 +46,8 @@ exports.signUp = async (req, res, next) => {
 
 }
 
-function generateAccessToken(id,name) {
-    return jwt.sign({ userId: id, name:name }, "secretkey");
+function generateAccessToken(id, name) {
+    return jwt.sign({ id: id, name: name }, "secretkey");
 }
 
 exports.signIn = async (req, res, next) => {
@@ -71,7 +74,6 @@ exports.signIn = async (req, res, next) => {
             else if (!response) {
                 return res.status(401).json({ customMessage: "User not authorized" });
             }
-
         })
 
     }
@@ -83,9 +85,8 @@ exports.signIn = async (req, res, next) => {
 
 exports.getAll = (req, res, next) => {
 
-    Expense.findAll({where:{userId:req.user.userId}})
+    Expense.findAll({ where: { userId: req.user.dataValues.id } })
         .then((data) => {
-            //console.log(req.user)
             res.status(201).json(data);
         })
         .catch(err => {
@@ -103,13 +104,14 @@ exports.addExpense = (req, res, next) => {
     if (!category || !description || !amount) {
         throw new Error("All fields are necessary");
     }
-     
-        Expense.create
+
+    req.user.createExpense
+        // Expense.create
         ({
             category: category,
             amount: amount,
-            description: description,
-            userId: req.user.userId
+            description: description
+            // userId: req.user.userId
         })
         .then((result) => {
             res.status(201).json(result);
@@ -123,7 +125,7 @@ exports.addExpense = (req, res, next) => {
 exports.deleteExpense = (req, res, next) => {
 
     const id = req.params.id
-    Expense.destroy({ where: { id: id } })
+    Expense.destroy({ where: { id: id, userId: req.user.dataValues.id } })
         .then(result => {
             res.sendStatus(201);
         })
@@ -157,7 +159,7 @@ exports.updateDetails = (req, res, next) => {
         throw new Error("All fields are necessary");
     }
 
-    Expense.findByPk(id)
+    Expense.findOne({ where: { id: id, userId: req.user.dataValues.id } })
         .then(async (result) => {
             result.category = category
             result.description = description
@@ -171,3 +173,99 @@ exports.updateDetails = (req, res, next) => {
         })
 }
 
+exports.buyPremium = (req, res, next) => {
+
+    const rzp = new Razorpay({
+        key_id: process.env.RAZORPAY_KEY_ID,
+        key_secret: process.env.RAZORPAY_KEY_SECRET
+    })
+
+    var options = {
+        amount: 3700,
+        currency: "INR"
+    }
+
+    rzp.orders.create(options, (err, order) => {
+        if (err) {
+            console.log(err);
+            throw new Error(JSON.stringify(err));
+        }
+        //console.log(order);
+        req.user.createOrder({ orderid: order.id, status: "PENDING" })
+            .then(() => {
+                return res.status(201).json({ order, key_id: rzp.key_id });
+            })
+            .catch(err => {
+                console.log(err);
+                return res.status(403).json({ message: "Something wrong", error: err });
+            })
+    })
+}
+
+exports.updatePremium = (req, res, next) => {
+
+    Order.findOne({ where: { orderid: req.body.order_id } })
+        .then(order => {
+            order.update({ paymentid: req.body.payment_id, status: "Success" })
+                .then(() => {
+                })
+                .catch(err => {
+                    console.log(err)
+                    res.status(500).json({ err: err });
+                })
+        })
+        .catch(err => {
+            console.log(err)
+            res.status(500).json({ err: err })
+            return
+        })
+
+    User.findByPk(req.user.dataValues.id)
+        .then(user => {
+            user.update({ premium: "yes" })
+                .then(() => {
+                    res.status(201).json({ message: "Tables Updated" })
+                })
+                .catch(err => {
+                    console.log(err)
+                    res.status(500).json({ err: err })
+                })
+
+        })
+        .catch(err => {
+            console.log(err)
+            res.status(500).json({ err: err })
+        })
+}
+
+exports.updateFailure=(req,res,next)=>{
+    
+    Order.findOne({where:{orderid:req.body.order_id}})
+    .then(order=>{
+        order.update({status:"failed"})
+        .then(()=>{
+            res.status(201).json({message:"Table updated"});
+        })
+        .catch(err=>{
+            console.log(err)
+            res.status(500).json(err);
+        })
+    })
+    .catch(err=>{
+        console.log(err);
+        res.status(500).json({err:err})
+    })
+}
+
+exports.checkPremium=(req,res,next)=>{
+    
+    const userId=req.user.dataValues.id;
+    User.findByPk(userId)
+    .then(user=>{
+        res.status(201).json({message:user.dataValues.premium});
+    })
+    .catch(err=>{
+        console.log(err);
+        res.status(500).json({message:"cannot fetch premium details"});
+    })
+}
